feat: 프로필 등록

+++ src/main/java/com/multipjt/multi_pjt/user/domain/mypage/ProfileResponseDTO.java
@@ -8,8 +8,8 @@ public class ProfileResponseDTO {
     private String profile_gender; // 성별
     private int profile_age; // 나이
     private String profile_region; //사는 지역
-    private float profile_weight; //몸무게
-    private float profile_height; //키
+    private double profile_weight; //몸무게
+    private double profile_height; //키
     private String profile_goal; //운동목표
     private int member_id; // 외래키
     private String profile_allergy; //알러지
